chore: simplify Dataset by moving low-level identity helpers into a separate module
