+++ test/datastore/dataset_test.rb
@@ -137,76 +137,6 @@ module Syskit::Log
                 end
             end
 
-            describe ".validate_encoded_short_digest" do
-                attr_reader :sha2
-                before do
-                    @sha2 = Digest::SHA2.hexdigest("TEST")
-                end
-                it "returns the digest unmodified if it is shorted than a full SHA2 digest" do
-                    assert_equal sha2[0..-2], Dataset.validate_encoded_short_digest(sha2[0..-2])
-                end
-                it "raises if the string is too long" do
-                    assert_raises(Dataset::InvalidDigest) do
-                        Dataset.validate_encoded_short_digest(sha2 + " ")
-                    end
-                end
-                it "raises if the string contains invalid characters for base64" do
-                    sha2[3, 1] = "_"
-                    assert_raises(Dataset::InvalidDigest) do
-                        Dataset.validate_encoded_short_digest(sha2)
-                    end
-                end
-                it "returns the digest unmodified if it is valid" do
-                    assert_equal sha2, Dataset.validate_encoded_short_digest(sha2)
-                end
-            end
-
-            describe ".valid_encoded_sha2?" do
-                attr_reader :sha2
-                before do
-                    @sha2 = Digest::SHA2.hexdigest("TEST")
-                end
-                it "returns false if the string is too short" do
-                    refute Dataset.valid_encoded_sha2?(sha2[0..-2])
-                end
-                it "raises if the string is too long" do
-                    refute Dataset.valid_encoded_sha2?(sha2 + " ")
-                end
-                it "raises if the string contains invalid characters for base64" do
-                    sha2[3, 1] = "_"
-                    refute Dataset.valid_encoded_sha2?(sha2)
-                end
-                it "returns true it is valid" do
-                    assert Dataset.valid_encoded_sha2?(sha2)
-                end
-            end
-
-            describe ".validate_encoded_sha2" do
-                attr_reader :sha2
-                before do
-                    @sha2 = Digest::SHA2.hexdigest("TEST")
-                end
-                it "raises if the string is too short" do
-                    assert_raises(Dataset::InvalidDigest) do
-                        Dataset.validate_encoded_sha2(sha2[0..-2])
-                    end
-                end
-                it "raises if the string is too long" do
-                    assert_raises(Dataset::InvalidDigest) do
-                        Dataset.validate_encoded_sha2(sha2 + " ")
-                    end
-                end
-                it "raises if the string contains invalid characters for base64" do
-                    sha2[3, 1] = "_"
-                    assert_raises(Dataset::InvalidDigest) do
-                        Dataset.validate_encoded_sha2(sha2)
-                    end
-                end
-                it "returns the digest unmodified if it is valid" do
-                    assert_equal sha2, Dataset.validate_encoded_sha2(sha2)
-                end
-            end
-
             describe "#compute_dataset_identity_from_files" do
                 it "returns a list of entries with full path, size and sha256 digest" do
                     roby_path = dataset_path + "roby-events.0.log"
@@ -369,7 +299,7 @@ module Syskit::Log
                     dataset.write_dataset_identity_to_metadata_file
                 end
                 it "computes a sha2 hash" do
-                    Dataset.validate_encoded_sha2(dataset.compute_dataset_digest)
+                    DatasetIdentity.validate_encoded_sha2(dataset.compute_dataset_digest)
                 end
                 it "is sensitive only to the file's relative paths" do
                     digest = dataset.compute_dataset_digest
